5е ДЗ. Стр. 9
